ais-ai/requirements.txt
anthropic==0.40.0
voyageai==0.2.3
qdrant-client==1.12.1
neo4j==5.26.0
grpcio==1.68.1
grpcio-tools==1.68.1
grpcio-reflection==1.68.1
protobuf==5.29.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
tiktoken==0.8.0
tree-sitter==0.21.3
tree-sitter-typescript==0.21.2
tree-sitter-javascript==0.21.4
tree-sitter-go==0.21.0
aiohttp==3.11.10
asyncio-throttle==1.0.2
